Extract contributor handling to a module
